필요없는 파일 삭제하고 용도가 비슷한 파일을 하나로 묶음
