+++ src/hooks/myInfo.js
@@ -1,9 +1,45 @@
 import { useEffect, useState } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 import { useRequestAuth } from './useRequestAuth';
 import { getApiEndpoint } from '../utils/util';
-import { useSaveUserInfo } from './userInfo';
+import { createReplacementUserAction } from '../redux/slice';
 
-function useRequestMyInfo() {
+export function useGetUserInfo() {
+  const { userInfo } = useSelector((state) => ({
+    userInfo: state.info.user,
+  }));
+
+  return {
+    userInfo: userInfo,
+  };
+}
+
+export function useSaveUserInfo() {
+  const dispatch = useDispatch();
+
+  const updateAll = ({ nickname, url, user_seq, field }) => {
+    dispatch(
+      createReplacementUserAction({
+        url,
+        field,
+        nickname,
+        user_seq,
+      })
+    );
+  };
+
+  const save = (data) => {
+    if (data) {
+      updateAll(data);
+    }
+  };
+
+  return {
+    save,
+  };
+}
+
+export function useMyInfo() {
   const { res, request } = useRequestAuth({
     endpoint: `${getApiEndpoint()}/me`,
     method: 'get',
@@ -32,5 +68,3 @@ function useRequestMyInfo() {
     userInfo,
   };
 }
-
-export default useRequestMyInfo;
